+++ hugr-core/src/extension/op_def.rs
@@ -794,7 +794,7 @@ pub(super) mod test {
             type Strategy = BoxedStrategy<Self>;
             fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
                 // TODO there is also  SignatureFunc::CustomFunc, but for now
-                // this is not serialised. When it is, we should generate
+                // this is not serialized. When it is, we should generate
                 // examples here .
                 any::<TypeSchemeRV>()
                     .prop_map(|x| SignatureFunc::TypeScheme(CustomValidator::from_polyfunc(x)))
@@ -807,7 +807,7 @@ pub(super) mod test {
             type Strategy = BoxedStrategy<Self>;
             fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
                 // TODO There is also LowerFunc::CustomFunc, but for now this is
-                // not serialised. When it is, we should generate examples here.
+                // not serialized. When it is, we should generate examples here.
                 any::<ExtensionSet>()
                     .prop_map(|extensions| LowerFunc::FixedHugr {
                         extensions,
@@ -841,7 +841,7 @@ pub(super) mod test {
                                 misc,
                                 signature_func,
                                 lower_funcs,
-                                // TODO ``constant_folder` is not serialised, we should
+                                // TODO ``constant_folder` is not serialized, we should
                                 // generate examples once it is.
                                 constant_folder: None,
                             })