+++ hugr-core/src/hugr.rs
@@ -51,7 +51,7 @@ pub struct Hugr {
 
 impl Default for Hugr {
     fn default() -> Self {
-        Self::new(crate::ops::Module)
+        Self::new(crate::ops::Module::new())
     }
 }
 
@@ -328,7 +328,7 @@ mod test {
 
     #[test]
     #[cfg_attr(miri, ignore)] // Opening files is not supported in (isolated) miri
-    #[ignore = "issue 1225: In serialisation we do not distinguish between unknown CustomConst serialised value invalid but known CustomConst serialised values"]
+    #[should_panic] // issue 1225: In serialization we do not distinguish between unknown CustomConst serialized value invalid but known CustomConst serialized values"
     fn hugr_validation_0() {
         // https://github.com/CQCL/hugr/issues/1091 bad case
         let mut hugr: Hugr = serde_json::from_reader(BufReader::new(