+++ hugr-passes/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.4.0 (2024-07-10)
+
+### Features
+
+- Add `force_order` pass. ([#1285](https://github.com/CQCL/hugr/pull/1285))
+
+### Refactor
+
+- [**breaking**] Remove `Value::Tuple` ([#1255](https://github.com/CQCL/hugr/pull/1255))
+
+
 ## 0.3.0 (2024-06-28)
 
 ### Features